requirements.txt
colorama>=0.4.6
fpdf2>=2.8.5
cloudscraper>=1.2.71
beautifulsoup4>=4.13.4
python-dotenv>=1.2.1
google-genai>=1.55.0
